Programmatically control grunt
Move grunt interface to a hook and control it via JS rather than shelling out.
Detect watch tasks and only allow sails to lift after the grunt tasks complete or a watch task is watching.
Add sails.config.grunt with options for tasks to run and path to gruntfile. Defaults to the same tasks and automatic gruntfile path detection as was previously the default.